Move UserTypes to shared
