+++ azext_iot/common/shared.py
@@ -304,3 +304,7 @@ class IotEdgeImageTermsCommands(Enum):
     Show = "show"
     Accept = "accept"
     Cancel = "cancel"
+
+class UserTypes(Enum):
+    user = "user"
+    servicePrincipal = "servicePrincipal"
