+++ azext_iot/tests/iothub/core/test_iothub_storage_int.py
@@ -11,8 +11,9 @@ from knack.util import CLIError
 from pathlib import Path
 
 from azext_iot.tests.iothub import IoTLiveScenarioTest
-from azext_iot.tests.settings import DynamoSettings, ENV_SET_TEST_IOTHUB_REQUIRED, ENV_SET_TEST_IOTHUB_OPTIONAL, UserTypes
-from azext_iot.common.utility import ensure_iothub_sdk_min_version, ensure_azure_namespace_path
+from azext_iot.tests.settings import DynamoSettings, ENV_SET_TEST_IOTHUB_REQUIRED, ENV_SET_TEST_IOTHUB_OPTIONAL
+from azext_iot.common.utility import ensure_iothub_sdk_min_version, 
+from azext_iot.common.shared import UserTypes
 
 from azext_iot.tests.generators import generate_generic_id
 # TODO: assert DEVICE_DEVICESCOPE_PREFIX format in parent device twin.
